Remove mentioning of VSCode integration
A VSCode extension was created for internal use for a company. However,
my desire to release, maintain and support this has disappeared.

This is mainly due to the low quality of tickets opened on the tracker.

- Many issues aren't related to preprocessing and are closed without
  action (eg. questions about bundlers)

- Some issues are related to user's environment and are closed /
  resolved without action

- Some issues are really just issues with Cypress, for which users
  mistankely direct here

- A large portion of the remaining issues simply doesn't contain a way
  of reproduction, which anyone within the QA space should know is a
  requirement

I have little faith in the user base's ability to produce good quality
tickets in the extension space, where environments varies wildly and
creating reproducible examples is a little more complex.

Furthermore, there is now an official extension for VSCode, which could
possibly be forked and / or modified to work well with Cypress +
Cucumber.